requirements.txt
pybind11
robotpy-ctre
wpilib
pyfrc
robotpy-pathfinder
scipy
numpy
